pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carla-driving-simulator-client"
version = "0.1.0"
description = "A comprehensive CARLA client for autonomous driving simulation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Akshay Chikhalkar", email = "akshaychikhalkar15@gmail.com"}
]
maintainers = [
    {name = "Akshay Chikhalkar", email = "akshaychikhalkar15@gmail.com"}
]
keywords = [
    "carla", "autonomous-driving", "simulation", "pygame", 
    "visualization", "scenario-testing", "vehicle-control"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
    "Topic :: Games/Entertainment :: Simulation",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Environment :: Web Environment",
]
requires-python = ">=3.11"
dependencies = [
    "pygame>=2.0.0",
    "numpy>=1.19.0",
    "matplotlib>=3.3.0",
    "tabulate>=0.8.7",
    "pyyaml>=5.4.0",
    "psycopg2-binary>=2.9.9",
    "SQLAlchemy>=2.0.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.10.0",
    "pytest-html>=3.2.0",
    "black>=21.5b2",
    "flake8>=3.9.0",
    "mypy>=0.812",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0",
    "myst-parser>=0.15.0",
]

[project.urls]
Homepage = "https://github.com/akshaychikhalkar/carla-driving-simulator-client"
Documentation = "https://carla-driving-simulator-client.readthedocs.io/"
Repository = "https://github.com/akshaychikhalkar/carla-driving-simulator-client"
"Bug Tracker" = "https://github.com/akshaychikhalkar/carla-driving-simulator-client/issues"

[project.scripts]
carla-simulator-client = "carla_simulator.cli:main"
csc = "carla_simulator.cli:main"

[tool.setuptools.packages.find]
include = ["carla_simulator*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pygame.*",
    "carla.*",
]
ignore_missing_imports = true
